rhiza-api/pyproject.toml
[project]
name = "rhiza"
version = "0.1.0"
description = ""
authors = [
    {name = "Steven Miklovic",email = "smiklov1@jh.edu"}
]
license = {text = "AGPL-3.0"}
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.118.0,<0.119.0)",
    "uvicorn (>=0.37.0,<0.38.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "google-generativeai (>=0.8.5,<0.9.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
